pyproject.toml
[tool.poetry]
name = "air-quality-forecast"
version = "0.1.0"
description = ""
authors = ["baniasbaabe <72874670+baniasbaabe@users.noreply.github.com>"]
readme = "README.md"
packages = [
    { include = "src" },
]


[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pre-commit = "^3.4.0"
loguru = "^0.7.2"
httpx = "^0.25.0"
hopsworks = "^3.2.0"
polars = "^0.19.8"
python-dotenv = "^1.0.0"
statsforecast = "^1.6.0"
utilsforecast = "^0.0.10"
comet-ml = "^3.34.1"
joblib = "^1.3.2"
pymongo = "^4.5.0"
pyyaml = "^6.0.1"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-randomly = "^3.15.0"
pytest-instafail = "^0.5.0"
pytest-cov = "^4.1.0"

[tool.poetry.group.linters]
optional = true

[tool.poetry.group.linters.dependencies]
mypy = "^1.5.1"
interrogate = "^1.5.0"
bandit = "^1.7.5"
ruff = "^0.0.292"

[tool.poetry.group.fixers]
optional = true

[tool.poetry.group.fixers.dependencies]
black = "^23.9.1"
docformatter = "^1.7.5"

[tool.ruff]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I"
]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.interrogate]
ignore-init-method = true
fail-under = 60

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
